fix(cli): honor SPECIFY_INIT_DIR for bundle commands
