+++ tests/test_init_dir_cli.py
@@ -91,6 +91,18 @@ def test_override_trailing_slash_tolerated(tmp_path, monkeypatch):
     assert "No workflows installed" in result.output
 
 
+def test_override_redirects_bundle_commands(tmp_path, monkeypatch):
+    web = _make_project(tmp_path, "web")
+    elsewhere = tmp_path / "elsewhere"
+    elsewhere.mkdir()
+    monkeypatch.chdir(elsewhere)
+    monkeypatch.setenv("SPECIFY_INIT_DIR", str(web))
+
+    result = runner.invoke(app, ["bundle", "list"])
+    assert result.exit_code == 0, result.output
+    assert "No bundles installed" in result.output
+
+
 def test_unset_override_uses_cwd(tmp_path, monkeypatch):
     """With SPECIFY_INIT_DIR unset, the project is the current directory."""
     cwd_proj = _make_project(tmp_path, "cwd")
@@ -124,6 +136,18 @@ def test_override_nonexistent_errors_no_fallback(tmp_path, monkeypatch):
     assert result.exit_code != 0
     assert "does not point to an existing directory" in result.output
     assert "No workflows installed" not in result.output  # no fallback to cwd
+
+
+def test_override_nonexistent_errors_bundle_commands_no_fallback(tmp_path, monkeypatch):
+    """Bundle commands also honor the strict override contract."""
+    cwd_proj = _make_project(tmp_path, "cwd")
+    monkeypatch.chdir(cwd_proj)
+    monkeypatch.setenv("SPECIFY_INIT_DIR", str(tmp_path / "does_not_exist"))
+
+    result = runner.invoke(app, ["bundle", "list"])
+    assert result.exit_code != 0
+    assert "does not point to an existing directory" in result.output
+    assert "No bundles installed" not in result.output
 
 
 def test_override_without_specify_errors_no_fallback(tmp_path, monkeypatch):
